Intern_Submissions/pravallika_tankala/requirements.txt
google-generativeai
python-dotenv
Django>=5.0
djangorestframework
streamlit
requests
sentence-transformers
faiss-cpu



